move infinite routes
