+++ frontends/infinite-corridor/src/infinite-pages/DemoPage.test.tsx
