+++ frontends/infinite-corridor/src/infinite-pages/HomePage.test.tsx
