+++ frontends/infinite-corridor/src/infinite-pages/HomePage.tsx
