+++ frontends/infinite-corridor/src/infinite-pages/SearchPage.test.tsx
