+++ frontends/infinite-corridor/src/infinite-pages/SearchPage.tsx
