+++ frontends/infinite-corridor/src/infinite-pages/field-details/EditFieldAppearanceForm.test.tsx
