+++ frontends/infinite-corridor/src/infinite-pages/field-details/EditFieldAppearanceForm.tsx
@@ -7,7 +7,7 @@ import * as Yup from "yup"
 import { FieldChannelAppearanceForm, useMutateField } from "../../api/fields"
 import type { FieldChannel } from "../../api/fields/interfaces"
 import FieldAvatar from "./components/FieldAvatar"
-import { makeFieldViewPath } from "../../pages/urls"
+import { makeFieldViewPath } from "../urls"
 
 type FormProps = {
   field: FieldChannel
