+++ frontends/infinite-corridor/src/infinite-pages/field-details/FieldAdminApp.tsx
