+++ frontends/infinite-corridor/src/infinite-pages/field-details/FieldPage.tsx
