+++ frontends/infinite-corridor/src/infinite-pages/field-details/FieldPageSkeleton.tsx
