+++ frontends/infinite-corridor/src/infinite-pages/field-details/WidgetsList.tsx
