+++ frontends/infinite-corridor/src/infinite-pages/field-details/components/FieldAvatar.test.tsx
