+++ frontends/infinite-corridor/src/infinite-pages/resource-lists/FavoritesPage.tsx
