+++ frontends/infinite-corridor/src/infinite-pages/resource-lists/ManageListDialogs.tsx
