+++ frontends/infinite-corridor/src/infinite-pages/resource-lists/ResourceListDetailsPage.test.tsx
