+++ frontends/infinite-corridor/src/pages/Home.tsx
@@ -1,0 +1,7 @@
+import React from "react"
+
+const HomePage: React.FC = () => {
+  return <div>Home</div>
+}
+
+export default HomePage
