+++ frontends/infinite-corridor/src/pages/errors/ForbiddenPage.test.tsx
@@ -1,7 +1,7 @@
 import React from "react"
 import { waitFor } from "@testing-library/react"
-import { renderWithProviders, screen } from "../test-utils"
-import { HOME } from "./urls"
+import { renderWithProviders, screen } from "../../test-utils"
+import { HOME } from "../urls"
 import ForbiddenPage from "./ForbiddenPage"
 
 test("The ForbiddenPage loads with meta", async () => {
