+++ frontends/infinite-corridor/src/pages/routes.ts
@@ -1,0 +1,12 @@
+import { RouteProps } from "react-router"
+import HomePage from "./Home"
+
+const routes: RouteProps[] = [
+  {
+    path:      "/",
+    component: HomePage,
+    exact:     true
+  }
+]
+
+export default routes
